+++ parlai/tasks/glue/README.md
@@ -1,0 +1,5 @@
+Task: Glue
+===============
+Description: GLUE, the General Language Understanding Evaluation benchmark is a collection of resources for training, evaluating, and analyzing natural language understanding systems. This task uses the `AbstractHuggingFaceTeacher` to load the dataset.
+
+Websites: https://huggingface.co/ and https://gluebenchmark.com/
