+++ parlai/tasks/glue/__init__.py
@@ -3,9 +3,3 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
-
-from parlai.utils.testing import AutoTeacherTest
-
-
-class TestDefaultTeacher(AutoTeacherTest):
-    task = 'huggingface'
