--- parlai/tasks/huggingface/test/huggingface_test.yml
@@ -1,43 +1,0 @@
-acts:
-- - episode_done: true
-    eval_labels:
-    - acceptable
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    text: Bill whistled past the house.
-- - episode_done: true
-    eval_labels:
-    - acceptable
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    text: The car honked its way down the road.
-- - episode_done: true
-    eval_labels:
-    - acceptable
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    text: Bill pushed Harry off the sofa.
-- - episode_done: true
-    eval_labels:
-    - acceptable
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    text: the kittens yawned awake and played.
-- - episode_done: true
-    eval_labels:
-    - acceptable
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    text: I demand that the more John eats, the more he pay.
-num_episodes: 1063
-num_examples: 1063
