--- parlai/tasks/huggingface/test/huggingface_train.yml
@@ -1,43 +1,0 @@
-acts:
-- - episode_done: true
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    labels:
-    - acceptable
-    text: Our friends won't buy this analysis, let alone the next one we propose.
-- - episode_done: true
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    labels:
-    - acceptable
-    text: One more pseudo generalization and I'm giving up.
-- - episode_done: true
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    labels:
-    - acceptable
-    text: One more pseudo generalization or I'm giving up.
-- - episode_done: true
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    labels:
-    - acceptable
-    text: The more we study verbs, the crazier they get.
-- - episode_done: true
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    labels:
-    - acceptable
-    text: Day by day the facts are getting murkier.
-num_episodes: 8551
-num_examples: 8551
