--- parlai/tasks/huggingface/test/huggingface_valid.yml
@@ -1,43 +1,0 @@
-acts:
-- - episode_done: true
-    eval_labels:
-    - acceptable
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    text: The sailors rode the breeze clear of the rocks.
-- - episode_done: true
-    eval_labels:
-    - acceptable
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    text: The weights made the rope stretch over the pulley.
-- - episode_done: true
-    eval_labels:
-    - acceptable
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    text: The mechanical doll wriggled itself loose.
-- - episode_done: true
-    eval_labels:
-    - acceptable
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    text: If you had eaten more, you would want less.
-- - episode_done: true
-    eval_labels:
-    - unacceptable
-    id: huggingface
-    label_candidates:
-    - unacceptable
-    - acceptable
-    text: As you eat the most, you want the least.
-num_episodes: 1043
-num_examples: 1043
